+++ src/Models/model.jl
@@ -1,9 +1,14 @@
-using Oceananigans.Grids: short_show
-using .TurbulenceClosures: ν₀, κ₀
-
-using Oceananigans.Buoyancy: validate_buoyancy
+using OrderedCollections: OrderedDict
 
 using Oceananigans.Fields
+
+using Oceananigans: AbstractArchitecture, CPU, GPU, ModelForcing
+using Oceananigans.Buoyancy: validate_buoyancy, SeawaterBuoyancy
+using Oceananigans.BoundaryConditions: HorizontallyPeriodicSolutionBCs, ModelBoundaryConditions
+using Oceananigans.TurbulenceClosures: ν₀, κ₀, ConstantIsotropicDiffusivity, TurbulentDiffusivities, with_tracers
+using Oceananigans: AbstractOutputWriter, AbstractDiagnostic
+using Oceananigans.Solvers: PoissonBCs, PoissonSolver
+using Oceananigans: TimeStepper
 
 mutable struct Model{TS, E, A<:AbstractArchitecture, G, T, B, R, SW, U, C, Φ, F,
                      BCS, S, K, OW, DI, Θ} <: AbstractModel
@@ -108,127 +113,4 @@ function Model(;
     return Model(architecture, grid, clock, buoyancy, coriolis, surface_waves, velocities, tracers,
                  pressures, forcing, closure, boundary_conditions, timestepper,
                  poisson_solver, diffusivities, output_writers, diagnostics, parameters)
-end
-
-"""Show the innards of a `Model` in the REPL."""
-Base.show(io::IO, model::Model) =
-    print(io, "Oceananigans.Model on a ", typeof(model.architecture), " architecture ",
-                                          "(time = ",  prettytime(model.clock.time),
-                                          ", iteration = ", model.clock.iteration, ") \n",
-              "├── grid: ", short_show(model.grid), '\n',
-              "├── tracers: ", tracernames(model.tracers), '\n',
-              "├── closure: ", typeof(model.closure), '\n',
-              "├── buoyancy: ", typeof(model.buoyancy), '\n',
-              "├── coriolis: ", typeof(model.coriolis), '\n',
-              "├── output writers: ", ordered_dict_show(model.output_writers, "│"), '\n',
-              "└── diagnostics: ", ordered_dict_show(model.diagnostics, " "))
-
-"""
-    ChannelModel(; kwargs...)
-
-Construct a `Model` with walls in the y-direction. This is done by imposing
-`FreeSlip` boundary conditions in the y-direction instead of `Periodic`.
-
-kwargs are passed to the regular `Model` constructor.
-"""
-ChannelModel(; boundary_conditions=ChannelSolutionBCs(), kwargs...) =
-    Model(; boundary_conditions=boundary_conditions, kwargs...)
-
-"""
-    NonDimensionalModel(; N, L, Re, Pr=0.7, Ro=Inf, float_type=Float64, kwargs...)
-
-Construct a "Non-dimensional" `Model` with resolution `N`, domain extent `L`,
-precision `float_type`, and the four non-dimensional numbers:
-
-    * `Re = U λ / ν` (Reynolds number)
-    * `Pr = U λ / κ` (Prandtl number)
-    * `Ro = U / f λ` (Rossby number)
-
-for characteristic velocity scale `U`, length-scale `λ`, viscosity `ν`,
-tracer diffusivity `κ`, and Coriolis parameter `f`. Buoyancy is scaled
-with `λ U²`, so that the Richardson number is `Ri=B`, where `B` is a
-non-dimensional buoyancy scale set by the user via initial conditions or
-forcing.
-
-Note that `N`, `L`, and `Re` are required.
-
-Additional `kwargs` are passed to the regular `Model` constructor.
-"""
-function NonDimensionalModel(; grid, float_type=Float64, Re, Pr=0.7, Ro=Inf,
-    buoyancy = BuoyancyTracer(),
-    coriolis = FPlane(float_type, f=1/Ro),
-     closure = ConstantIsotropicDiffusivity(float_type, ν=1/Re, κ=1/(Pr*Re)),
-    kwargs...)
-
-    return Model(; float_type=float_type, grid=grid, closure=closure,
-                   coriolis=coriolis, tracers=(:b,), buoyancy=buoyancy, kwargs...)
-end
-
-#####
-##### Utils
-#####
-
-float_type(m::AbstractModel) = eltype(model.grid)
-
-"""
-    VelocityFields(arch, grid)
-
-Return a NamedTuple with fields `u`, `v`, `w` initialized on
-the architecture `arch` and `grid`.
-"""
-function VelocityFields(arch, grid)
-    u = FaceFieldX(arch, grid)
-    v = FaceFieldY(arch, grid)
-    w = FaceFieldZ(arch, grid)
-    return (u=u, v=v, w=w)
-end
-
-"""
-    TracerFields(arch, grid)
-
-Return a NamedTuple with tracer fields initialized
-as `CellField`s on the architecture `arch` and `grid`.
-"""
-function TracerFields(arch, grid, tracernames)
-    tracerfields = Tuple(CellField(arch, grid) for c in tracernames)
-    return NamedTuple{tracernames}(tracerfields)
-end
-
-TracerFields(arch, grid, ::Union{Tuple{}, Nothing}) = NamedTuple{()}(())
-TracerFields(arch, grid, tracer::Symbol) = TracerFields(arch, grid, tuple(tracer))
-TracerFields(arch, grid, tracers::NamedTuple) = tracers
-
-tracernames(::Nothing) = ()
-tracernames(name::Symbol) = tuple(name)
-tracernames(names::NTuple{N, Symbol}) where N = :u ∈ names ? names[4:end] : names
-tracernames(::NamedTuple{names}) where names = tracernames(names)
-
-"""
-    PressureFields(arch, grid)
-
-Return a NamedTuple with pressure fields `pHY′` and `pNHS`
-initialized as `CellField`s on the architecture `arch` and `grid`.
-"""
-function PressureFields(arch, grid)
-    pHY′ = CellField(arch, grid)
-    pNHS = CellField(arch, grid)
-    return (pHY′=pHY′, pNHS=pNHS)
-end
-
-"""
-    Tendencies(arch, grid, tracernames)
-
-Return a NamedTuple with tendencies for all solution fields
-(velocity fields and tracer fields), initialized on
-the architecture `arch` and `grid`.
-"""
-function Tendencies(arch, grid, tracernames)
-
-    velocities = (u = FaceFieldX(arch, grid),
-                  v = FaceFieldY(arch, grid),
-                  w = FaceFieldZ(arch, grid))
-
-    tracers = TracerFields(arch, grid, tracernames)
-
-    return merge(velocities, tracers)
 end
