+++ src/Models/model_utils.jl
@@ -1,0 +1,64 @@
+float_type(m::AbstractModel) = eltype(model.grid)
+
+"""
+    VelocityFields(arch, grid)
+
+Return a NamedTuple with fields `u`, `v`, `w` initialized on
+the architecture `arch` and `grid`.
+"""
+function VelocityFields(arch, grid)
+    u = FaceFieldX(arch, grid)
+    v = FaceFieldY(arch, grid)
+    w = FaceFieldZ(arch, grid)
+    return (u=u, v=v, w=w)
+end
+
+"""
+    TracerFields(arch, grid)
+
+Return a NamedTuple with tracer fields initialized
+as `CellField`s on the architecture `arch` and `grid`.
+"""
+function TracerFields(arch, grid, tracernames)
+    tracerfields = Tuple(CellField(arch, grid) for c in tracernames)
+    return NamedTuple{tracernames}(tracerfields)
+end
+
+TracerFields(arch, grid, ::Union{Tuple{}, Nothing}) = NamedTuple{()}(())
+TracerFields(arch, grid, tracer::Symbol) = TracerFields(arch, grid, tuple(tracer))
+TracerFields(arch, grid, tracers::NamedTuple) = tracers
+
+tracernames(::Nothing) = ()
+tracernames(name::Symbol) = tuple(name)
+tracernames(names::NTuple{N, Symbol}) where N = :u ∈ names ? names[4:end] : names
+tracernames(::NamedTuple{names}) where names = tracernames(names)
+
+"""
+    PressureFields(arch, grid)
+
+Return a NamedTuple with pressure fields `pHY′` and `pNHS`
+initialized as `CellField`s on the architecture `arch` and `grid`.
+"""
+function PressureFields(arch, grid)
+    pHY′ = CellField(arch, grid)
+    pNHS = CellField(arch, grid)
+    return (pHY′=pHY′, pNHS=pNHS)
+end
+
+"""
+    Tendencies(arch, grid, tracernames)
+
+Return a NamedTuple with tendencies for all solution fields
+(velocity fields and tracer fields), initialized on
+the architecture `arch` and `grid`.
+"""
+function Tendencies(arch, grid, tracernames)
+
+    velocities = (u = FaceFieldX(arch, grid),
+                  v = FaceFieldY(arch, grid),
+                  w = FaceFieldZ(arch, grid))
+
+    tracers = TracerFields(arch, grid, tracernames)
+
+    return merge(velocities, tracers)
+end
