+++ src/utils.jl
@@ -1,3 +1,5 @@
+using OrderedCollections: OrderedDict
+
 #####
 ##### Adapting structures to be able to pass them to GPU CUDA kernels.
 #####
@@ -145,7 +147,7 @@ milliseconds (ms), seconds (s), minutes (min), hours (hr), or days (day).
 function prettytime(t)
     # Modified from: https://github.com/JuliaCI/BenchmarkTools.jl/blob/master/src/trials.jl
     iszero(t) && return "0.000 s"
-    
+
     if t < 1e-6
         value, units = t * 1e9, "ns"
     elseif t < 1e-3
